+++ PADO/PADO/iOS/Surfing/View/SurfingView.swift
@@ -147,104 +147,4 @@ struct SurfingView: View {
     }
 }
 
-// PHPickerViewController를 사용하는 SwiftUI 뷰
-struct PhotoPicker: UIViewControllerRepresentable {
-    @Binding var pickerResult: [PHPickerResult]
-    @Binding var selectedImage: UIImage?
-    // Image 타입으로 변환된 이미지를 위한 새로운 바인딩
-    @Binding var selectedSwiftUIImage: Image
-    
-    func makeUIViewController(context: Context) -> PHPickerViewController {
-        var config = PHPickerConfiguration(photoLibrary: .shared())
-        config.filter = .images
-        config.selectionLimit = 1 // 0으로 설정하면 무제한 선택 가능
-        config.disabledCapabilities = .selectionActions
-        
-        let picker = PHPickerViewController(configuration: config)
-        picker.delegate = context.coordinator
-        return picker
-    }
-    
-    func updateUIViewController(_ uiViewController: PHPickerViewController, context: Context) {
-        
-    }
-    
-    func makeCoordinator() -> Coordinator {
-        Coordinator(self)
-    }
-    
-    class Coordinator: NSObject, PHPickerViewControllerDelegate {
-        var parent: PhotoPicker
-        
-        init(_ parent: PhotoPicker) {
-            self.parent = parent
-        }
-        
-        // PHPickerViewController에서 사진을 선택한 후 호출됩니다.
-        func picker(_ picker: PHPickerViewController, didFinishPicking results: [PHPickerResult]) {
-            parent.pickerResult = results
-            
-            guard !results.isEmpty else { return }
-            let itemProvider = results[0].itemProvider
-            
-            // 선택된 사진의 데이터를 가져옵니다.
-            if itemProvider.canLoadObject(ofClass: UIImage.self) {
-                itemProvider.loadObject(ofClass: UIImage.self) { image, error in
-                    DispatchQueue.main.async {
-                        if let image = image as? UIImage {
-                            self.parent.selectedImage = image
-                            // UIImage를 Image로 변환
-                            self.parent.selectedSwiftUIImage = Image(uiImage: image)
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
 
-struct CameraAccessView: UIViewControllerRepresentable {
-    @Binding var isShown: Bool
-    @Binding var myimage: Image
-    @Binding var mysourceType: UIImagePickerController.SourceType
-    
-    func updateUIViewController(_ uiViewController: UIImagePickerController, context: UIViewControllerRepresentableContext<CameraAccessView>) {
-        
-    }
-    
-    func makeUIViewController(context: UIViewControllerRepresentableContext<CameraAccessView>) -> UIImagePickerController {
-        let obj = UIImagePickerController()
-        obj.sourceType = mysourceType
-        obj.delegate = context.coordinator
-        return obj
-    }
-    
-    func makeCoordinator() -> CameraCoordinator {
-        return CameraCoordinator(isShown: $isShown, myimage: $myimage)
-    }
-}
-
-// UIImagePickerController와의 상호작용을 처리하기 위한 코디네이터 클래스
-class CameraCoordinator: NSObject, UIImagePickerControllerDelegate, UINavigationControllerDelegate {
-    @Binding var isShown: Bool
-    @Binding var myimage: Image
-    
-    init(isShown: Binding<Bool>, myimage: Binding<Image>) {
-        _isShown = isShown
-        _myimage = myimage
-    }
-    
-    // 이미지 선택이 완료되었을때 호출
-    func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
-        if let image = info[UIImagePickerController.InfoKey.originalImage] as? UIImage {
-            print(image)
-            myimage = Image.init(uiImage: image)
-        }
-        isShown = false
-    }
-    
-    // 이미지 선택이 취소 되었을 때 호출
-    func imagePickerControllerDidCancel(_ picker: UIImagePickerController) {
-        isShown = false
-    }
-}
